web/backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.6
py2neo==2021.2.3
sqlparse==0.4.3
